Refactor. 레이아웃 컴포넌트를 레이아웃 폴더로 옮김
